+++ src/components/improvements.js
@@ -1,5 +1,4 @@
 var Prefs = require('../preferences'),
-    MenuActions = require('../menu-actions'),
     Script = require('../script'),
     Page = require('../page');
 
@@ -37,47 +36,6 @@ function addMorePopovers(more) {
     });
 }
 
-function addDupeCheck() {
-    function addDupeWarn(historybtn, dupe) {
-        historybtn.removeClass('btn-default').addClass(dupe ? 'btn-danger' : 'btn-success');
-    }
-
-    function checkDuped(oid) {
-        $.get("/item/" + oid, function (html) {
-            var dupe = /Refer to entries in the item history <strong>where the item ID is not chronological/.test(html);
-            window.dupeCache[oid] = dupe;
-            // Use the newest history button in case user hovers away
-            window.addDupeWarn($('.popover .fa-calendar-o').parent(), dupe);
-        });
-    }
-
-    function createDetails(item) {
-        var details = window.dupe_createDetails(item),
-            oid = item.attr('data-original-id'),
-            historybtn = details.find('.fa-calendar-o').parent();
-
-        if (window.dupeCache[oid] != null) { // undefined/null (none/in progress)
-            window.addDupeWarn(historybtn, window.dupeCache[oid]);
-        } else {
-            historybtn.mouseover(function () {
-                if (window.dupeCache[oid] !== null) { // not in progress
-                    window.dupeCache[oid] = null;
-                    setTimeout(function () {
-                        window.checkDuped(oid);
-                    }, 80);
-                }
-            });
-        }
-
-        return details;
-    }
-
-    Script.exec('var dupe_createDetails = window.createDetails, dupeCache = {};'+
-                'window.checkDuped = ' + checkDuped + ';'+
-                'window.addDupeWarn = ' + addDupeWarn + ';'+
-                'window.createDetails = ' + createDetails + ';');
-}
-
 function global() {
     var account = $('#profile-dropdown-container a[href="/my/account"]'),
         help = $('.dropdown a[href="/help"]'),
@@ -86,8 +44,6 @@ function global() {
     if (account.length) account.parent().after('<li><a href="/my/preferences"><i class="fa fa-fw fa-cog"></i> My Preferences</a></li>');
     if (help.length) help.parent().before('<li><a href="/lotto"><i class="fa fa-fw fa-money"></i> Lotto</a></li>');
     if (more.length) addMorePopovers(more);
-
-    addDupeCheck();
 }
 
 function index() {
@@ -118,7 +74,9 @@ function index() {
             $("#notifications_new").remove();
 
             updateNotifications();
-        } else if (updatec === 'click') {
+        }
+
+        if (updatec === 'click' || updatec === 'listing') {
             notifs.find(".notification").click(updateNotifications);
         }
     }