+++ suite.meta.js
@@ -3,7 +3,7 @@
 // @name         backpack.tf enhancement suite
 // @namespace    http://steamcommunity.com/id/caresx/
 // @author       cares
-// @version      1.1.5
+// @version      1.2.0
 // @description  Enhances your backpack.tf experience.
 // @match        *://backpack.tf/*
 // @match        *://lv*.backpack.tf/*
@@ -11,7 +11,7 @@
 // @match        *://csgo.backpack.tf/*
 // @require      https://code.jquery.com/jquery-2.1.3.min.js
 // @require      https://maxcdn.bootstrapcdn.com/bootstrap/3.3.4/js/bootstrap.min.js
-// @require      https://cdn.rawgit.com/caresx/steam-econcc/2551ce827114e8fd11e94feaa9681bf4aa302379/econcc.js
+// @require      https://cdn.rawgit.com/caresx/steam-econcc/d3762b8978616b7f1e2086deb8fd62552b0746c0/econcc.js
 // @require      https://cdn.rawgit.com/visionmedia/page.js/1102353025a984f7c88f38538bd62ff89e1eeee6/page.js
 // @require      https://cdnjs.cloudflare.com/ajax/libs/moment.js/2.10.2/moment.min.js
 // @downloadURL  https://caresx.github.io/backpacktf-enhancement-suite/suite.user.js